+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - [#66](https://github.com/Kashoo/synctos/issues/66): Modular document definition files
 - [#69](https://github.com/Kashoo/synctos/issues/69): Helper function to determine whether a document is missing or deleted
 - [#72](https://github.com/Kashoo/synctos/issues/72): New property validation type for type identifier properties
+- [#73](https://github.com/Kashoo/synctos/issues/73): Include an implicit type property when a simple type filter is used
 
 ## [1.5.0] - 2016-12-14
 ### Added